requirements.txt
-e .
ipython>=9.3.0
jupyterlab>=4.0.0

# Data visualization
matplotlib>=3.7.0

# Documentation
mkdocs>=1.5.0
notebook>=7.0.0
numpy>=2.3.0
# Core data processing
pandas>=2.3.0

# Development setup
pip>=23.0.0

# Development and testing
pytest>=8.4.0
python-dotenv>=1.0.0
requests>=2.32.4

# Code quality
ruff>=0.1.0

# Machine learning (optional)
scikit-learn>=1.3.0

# Database and API clients
supabase>=2.15.3
